requirements.txt
keras==2.13.1
matplotlib==3.7.3
numpy==1.24.3
opencv-python==4.8.1.78
pandas==2.0.3
Pillow==10.0.1
scikit-learn==1.3.1
scipy==1.10.1
streamlit==1.27.2
tensorflow==2.13.0

